Create Wholesaler/Inventory action

+++ Application/DTOs/WholesalerInventoryDto.cs
@@ -3,5 +3,5 @@ public class WholesalerInventoryDto
 {
     public string BeerId { get; set; } = default!;
     public int Quantity { get; set; }
-    public int FixedPrice { get; set; }
+    public decimal FixedPrice { get; set; }
 }
